fix tcr weight

+++ parsing.py
@@ -11,5 +11,5 @@ def get_parser():
     parser.add_argument('--checkpoint', type=str, default='./checkpoint/model_checkpoint.pt', help='Model checkpoint')
     parser.add_argument('--workers', type=int, default=4, help='number of workers for data loader to use')
     #parser.add_argument('--seed', type=int, default=123, help='random seed to use. Default=123')
-    parser.add_argument('--weight_tcr', type=int, default=0.01, help='random seed to use. Default=123')
+    parser.add_argument('--weight_tcr', type=float, default=0.01, help='random seed to use. Default=123')
     return parser
